feat: Add book management features (add book, list books, and database integration)

+++ app/src/main/java/com/omeroztoprak/booktracker/DatabaseHelper.java
@@ -1,5 +1,6 @@
 package com.omeroztoprak.booktracker;
 
+import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
@@ -8,12 +9,19 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static final String DATABASE_NAME = "BookTrackerDB";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
 
     public static final String TABLE_USERS = "Users";
     public static final String COLUMN_ID = "id";
     public static final String COLUMN_USERNAME = "username";
     public static final String COLUMN_PASSWORD = "password";
+
+    public static final String TABLE_BOOKS = "Books";
+    public static final String COLUMN_BOOK_ID = "book_id";
+    public static final String COLUMN_BOOK_TITLE = "title";
+    public static final String COLUMN_BOOK_AUTHOR = "author";
+    public static final String COLUMN_BOOK_CATEGORY = "category";
+    public static final String COLUMN_BOOK_COMMENT = "comment";
 
     public DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -27,12 +35,24 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 + COLUMN_PASSWORD + " TEXT"
                 + ")";
         db.execSQL(CREATE_USERS_TABLE);
+
+        String CREATE_BOOKS_TABLE = "CREATE TABLE " + TABLE_BOOKS + "("
+                + COLUMN_BOOK_ID + " INTEGER PRIMARY KEY AUTOINCREMENT,"
+                + COLUMN_BOOK_TITLE + " TEXT,"
+                + COLUMN_BOOK_AUTHOR + " TEXT,"
+                + COLUMN_BOOK_CATEGORY + " TEXT,"
+                + COLUMN_BOOK_COMMENT + " TEXT"
+                + ")";
+        db.execSQL(CREATE_BOOKS_TABLE);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
-        onCreate(db);
+        if (oldVersion < 2) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_BOOKS);
+            onCreate(db);
+        }
     }
 
     public boolean checkUserCredentials(String username, String password) {
@@ -51,9 +71,32 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public boolean addUser(String username, String password) {
         SQLiteDatabase db = this.getWritableDatabase();
-        String query = "INSERT INTO " + TABLE_USERS + "(" + COLUMN_USERNAME + ", " + COLUMN_PASSWORD + ")"
-                + " VALUES (?, ?)";
-        db.execSQL(query, new String[]{username, password});
-        return true;
+        ContentValues values = new ContentValues();
+        values.put(COLUMN_USERNAME, username);
+        values.put(COLUMN_PASSWORD, password);
+
+        long result = db.insert(TABLE_USERS, null, values);
+        return result != -1;
+    }
+
+    public boolean addBook(String title, String author, String category, String comment) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        ContentValues values = new ContentValues();
+        values.put(COLUMN_BOOK_TITLE, title);
+        values.put(COLUMN_BOOK_AUTHOR, author);
+        values.put(COLUMN_BOOK_CATEGORY, category);
+        values.put(COLUMN_BOOK_COMMENT, comment);
+
+        long result = db.insert(TABLE_BOOKS, null, values);
+        return result != -1;
+    }
+
+    public Cursor getAllBooks() {
+        SQLiteDatabase db = this.getReadableDatabase();
+        return db.rawQuery("SELECT "
+                + COLUMN_BOOK_ID + " as _id, "
+                + COLUMN_BOOK_TITLE + ", "
+                + COLUMN_BOOK_AUTHOR + " FROM "
+                + TABLE_BOOKS, null);
     }
 }
